lesson21/money/pyproject.toml
[tool.poetry]
name = "money"
version = "0.1.0"
description = ""
authors = ["Vladlen Papulov <vladlen.papulov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^1.1.2"
Flask-SQLAlchemy = "^2.4.4"
Flask-Migrate = "^2.5.3"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
